pyproject.toml
[project]
name = "livekit-telephony-sts"
version = "0.1.0"
description = "LiveKit Telephony STS"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "livekit>=0.22.0",
    "livekit-agents>=0.12.18",
    "livekit-plugins-openai>=0.12.2",
    "livekit-plugins-silero>=0.7.5",
    "python-dotenv>=1.1.0",
]
